Move route writer to own module and only write when pages change
